+++ logger/token.json
@@ -1,0 +1,3 @@
+{
+  "api_token": "eyJhcGlfYWRkcmVzcyI6Imh0dHBzOi8vYXBwLm5lcHR1bmUuYWkiLCJhcGlfdXJsIjoiaHR0cHM6Ly9hcHAubmVwdHVuZS5haSIsImFwaV9rZXkiOiIwYzFhNGNlZi1jOTY4LTQ4YmEtODk2Ni1kY2YyMDg4MGYzM2EifQ== "
+}
