--- utils/logger/token.json
@@ -1,3 +1,0 @@
-{
-  "api_token": "enter_your_token"
-}
